feat: add Time class with class constant example
Added a 'Time' class to demonstrate the use of class constants in OOP. Defined a constant 'DAY_IN_SECONDS' to represent the number of seconds in a day. Included a method 'tomorrow' that uses the class constant to calculate the time for the next day. Showed how to access the constant using 'self::'.