feat: add clone method to Product class
Added a clone method (__clone) to the 'Product' class. This magic method is executed when an object is cloned. Explained that the cloned object will have the same properties and methods as the original object but will have a different memory address. Included a construct method to demonstrate object creation.