feat: add explanation for autoload method
Added explanations for the autoload method in OOP. Described how the autoload method is executed when a class has not been declared. Explained the use of the __autoload() function to search for undefined classes in included files and how it is called outside of a class to handle missing class definitions.