feat: add explanation for PHP namespaces
Added explanations for PHP namespaces. Described how namespaces are used to avoid name collisions between classes, functions, and constants. Explained the declaration of namespaces at the beginning of the file, the use of class aliases, and the role of Composer as a package manager for PHP. Included an example namespace 'App\Library' and the use of an alias for 'Illuminate\Support\Facades\Facade'.